+++ .gitignore
@@ -61,6 +61,9 @@ coverage.xml
 .pytest_cache/
 cover/
 
+# Pytest_file
+.pytest
+
 # Translations
 *.mo
 *.pot
