+++ test/test_delete.py
@@ -1,0 +1,2 @@
+from tinkermodellor import TinkerModellor as tkm
+
